+++ .github/workflows/rust.yml
@@ -58,7 +58,7 @@ jobs:
       - name: Build
         run: cargo  build --verbose
       - name: Run tests
-        run: cargo test --verbose --no-default-features
+        run: cargo test --verbose --no-default-features --features=sabiork 
   check:
     name: Code coverage
     runs-on: ubuntu-latest
