+++ Cargo.toml
@@ -29,3 +29,4 @@ itertools = "0.10"
 extension-module = ["pyo3/extension-module"]
 default = ["extension-module"]
 dependencies = ["pyo3"]
+sabiork = []
